# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * APEX TRAIL — E-commerce Product Catalog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductSpecs = {[key: string]: string};

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  categorySlug: string;
  images: string[];
  description: string;
  shortDescription?: string;
  rating: number;
  reviewCount: number;
  stock: number;
  tags: string[];
  createdAt: string;
  featured: boolean;
  specs?: ProductSpecs;
}

export interface ProductPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  productCount: number;
  imageUrl: string;
}

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  avgRating: number;
  newArrivalsCount: number;
  inStockCount: number;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  subject?: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactAck {
  success: boolean;
  message: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
sort?: ListProductsSort;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
tags?: string;
page?: number;
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
  'rating-desc': 'rating-desc',
} as const;