# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: APEX TRAIL — E-commerce Product Catalog API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: catalog
    description: Catalog summary and discovery
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filtering, sorting, and pagination
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price-asc, price-desc, rating-desc]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: tags
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: Get featured/hero products for the homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products for a product detail page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories with counts
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found

  /catalog/summary:
    get:
      operationId: getCatalogSummary
      tags: [catalog]
      summary: Homepage stats — total products, categories, avg rating, new arrivals count
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogSummary"

  /catalog/new-arrivals:
    get:
      operationId: listNewArrivals
      tags: [catalog]
      summary: Products added in the last 30 days
      responses:
        "200":
          description: New arrivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /catalog/bestsellers:
    get:
      operationId: listBestsellers
      tags: [catalog]
      summary: Top-rated products sorted by rating and review count
      responses:
        "200":
          description: Bestseller products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /catalog/tags:
    get:
      operationId: listTags
      tags: [catalog]
      summary: All unique product tags for faceted filtering
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /contact:
    post:
      operationId: submitContact
      tags: [catalog]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactAck"
        "422":
          description: Validation error

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Product:
      type: object
      required: [id, name, slug, price, category, categorySlug, images, description, rating, reviewCount, stock, tags, createdAt, featured]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        categorySlug:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        shortDescription:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        featured:
          type: boolean
        specs:
          type: object
          additionalProperties:
            type: string

    ProductPage:
      type: object
      required: [products, total, page, limit, totalPages]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    Category:
      type: object
      required: [id, name, slug, description, productCount, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        productCount:
          type: integer
        imageUrl:
          type: string

    CatalogSummary:
      type: object
      required: [totalProducts, totalCategories, avgRating, newArrivalsCount, inStockCount]
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        avgRating:
          type: number
        newArrivalsCount:
          type: integer
        inStockCount:
          type: integer

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        subject:
          type: string
        message:
          type: string
          minLength: 10

    ContactAck:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string