# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * APEX TRAIL — E-commerce Product Catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  avgRating: number;
  newArrivalsCount: number;
  inStockCount: number;
}