# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * APEX TRAIL — E-commerce Product Catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  productCount: number;
  imageUrl: string;
}