# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * APEX TRAIL — E-commerce Product Catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecs } from './productSpecs';

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  categorySlug: string;
  images: string[];
  description: string;
  shortDescription?: string;
  rating: number;
  reviewCount: number;
  stock: number;
  tags: string[];
  createdAt: string;
  featured: boolean;
  specs?: ProductSpecs;
}