# lib/api-zod/src/generated/types/productPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * APEX TRAIL — E-commerce Product Catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}